+++ packages/tideline/js/plot/timechange.js
@@ -16,12 +16,9 @@
  */
 
 import _ from 'lodash'
-import i18next from 'i18next'
-import moment from 'moment-timezone'
 import * as d3 from 'd3'
-
-import * as constants from '../data/util/constants'
 import timeChangeImage from '../../img/timechange/timechange.svg'
+import utils from './util/utils'
 
 /**
  * Default configuration for this component
@@ -54,6 +51,7 @@ function plotTimeChange(pool, opts = {}) {
         .append('g')
         .classed('d3-timechange-group', true)
         .attr('id', d => `timechange_${d.id}`)
+        .attr('data-testid', d => `timechange_${d.id}`)
 
       timechange.addTimeChangeToPool(timechangeGroup)
 
@@ -73,78 +71,15 @@ function plotTimeChange(pool, opts = {}) {
       .attr('width', opts.size)
       .attr('height', opts.size)
 
-    selection.on('mouseover', timechange.displayTooltip)
-    selection.on('mouseout', timechange.removeTooltip)
-  }
-
-  timechange.removeTooltip = (event, d) => {
-    d3.select(`#tooltip_${d.id}`).remove()
-  }
-
-  timechange.displayTooltip = (event, d) => {
-    const t = i18next.t.bind(i18next)
-    const mFrom = moment.tz(d.from.time, d.from.timeZoneName)
-    const mTo = moment.tz(d.to.time, d.to.timeZoneName)
-
-    let format = 'h:mm a'
-    if (mFrom.year() !== mTo.year()) {
-      format = constants.dateTimeFormats.MMM_D_YYYY_H_MM_A_FORMAT
-    } else if (mFrom.month() !== mTo.month()) {
-      format = constants.dateTimeFormats.MMM_D_H_MM_A_FORMAT
-    } else if (mFrom.date() !== mTo.date()) {
-      format = constants.dateTimeFormats.DDDD_H_MM_A
-    } else {
-      format = constants.dateTimeFormats.H_MM_A_FORMAT
-    }
-
-    const fromDate = mFrom.format(format)
-    const toDate = mTo.format(format)
-
-    let changeType
-    let tzLine1 = null
-    let tzLine2 = null
-
-    if (d.from.timeZoneName === d.to.timeZoneName) {
-      changeType = t('Time Change')
-      tzLine1 = `<span class="fromto">${t('from')}</span> ${fromDate} <span class="fromto">${t('to')}</span> ${toDate}`
-    } else {
-      changeType = t('Timezone Change')
-      tzLine1 = `<span class="fromto">${t('from')}</span> ${fromDate} - ${d.from.timeZoneName}`
-      tzLine2 = `<span class="fromto">${t('to')}</span> ${toDate} - ${d.to.timeZoneName}`
-    }
-
-    const tooltips = pool.tooltips()
-    const tooltip = tooltips.addForeignObjTooltip({
-      cssClass: 'svg-tooltip-timechange',
-      datum: d,
-      shape: 'generic',
-      xPosition: timechange.xPositionCenter,
-      yPosition: timechange.yPositionCenter
+    selection.on('mouseover', function (event, d) {
+      opts.onTimeChangeHover({
+        data: d,
+        rect: utils.getTooltipContainer(this)
+      })
     })
 
-    const { foGroup } = tooltip
-    foGroup.append('p').append('span').attr('class', 'secondary').html(tzLine1)
-    if (tzLine2) {
-      foGroup.append('p').append('span').attr('class', 'secondary').html(tzLine2)
-    }
-    foGroup.append('p').append('span').attr('class', 'mainText').html(changeType)
-
-    const dims = tooltips.foreignObjDimensions(foGroup)
-
-    // foGroup.node().parentNode is the <foreignObject> itself
-    // because foGroup is actually the top-level <xhtml:div> element
-    tooltips.anchorForeignObj(d3.select(foGroup.node().parentNode), {
-      w: dims.width + opts.tooltipPadding,
-      h: dims.height,
-      x: timechange.xPositionCenter(d),
-      y: -dims.height,
-      orientation: {
-        default: 'leftAndDown',
-        leftEdge: 'rightAndDown',
-        rightEdge: 'leftAndDown'
-      },
-      shape: 'generic',
-      edge: tooltip.edge
+    selection.on('mouseout', function () {
+      opts.onTimeChangeOut()
     })
   }